move algorithm into core
