+++ core/include/traccc/seeding/seeding_algorithm.hpp
