+++ examples/CMakeLists.txt
@@ -1,3 +1,2 @@
-add_subdirectory(algorithms)
 add_subdirectory(performance)
 add_subdirectory(run)
