--- examples/algorithms/CMakeLists.txt
@@ -1,9 +1,0 @@
-add_subdirectory(cpu)
-
-if (TRACCC_BUILD_CUDA)
-  add_subdirectory(cuda)
-endif()
-
-if (TRACCC_BUILD_SYCL)
-  add_subdirectory(sycl)
-endif()
