+++ examples/run/cpu/CMakeLists.txt
@@ -5,12 +5,12 @@
 # Mozilla Public License Version 2.0
 
 traccc_add_executable( seeding_example "seeding_example.cpp"
-   LINK_LIBRARIES vecmem::core traccc::core traccc::algorithms traccc::io 
+   LINK_LIBRARIES vecmem::core traccc::core traccc::io 
    traccc::performance Boost::program_options)
 
 traccc_add_executable( seq_example "seq_example.cpp"
-   LINK_LIBRARIES vecmem::core traccc::core traccc::algorithms traccc::io 
+   LINK_LIBRARIES vecmem::core traccc::core traccc::io 
    traccc::performance Boost::program_options)
 
 traccc_add_executable( ccl_example "ccl_example.cpp"
-   LINK_LIBRARIES vecmem::core traccc::core traccc::algorithms traccc::io)
+   LINK_LIBRARIES vecmem::core traccc::core traccc::io)
