+++ examples/run/cpu/seeding_example.cpp
@@ -11,8 +11,8 @@
 #include "traccc/io/writer.hpp"
 
 // algorithms
+#include "traccc/seeding/seeding_algorithm.hpp"
 #include "traccc/seeding/track_params_estimation.hpp"
-#include "traccc/track_finding/seeding_algorithm.hpp"
 
 // performance
 #include "traccc/efficiency/seeding_performance_writer.hpp"
