+++ examples/run/cuda/CMakeLists.txt
@@ -8,10 +8,8 @@
 include( traccc-compiler-options-cuda )
 
 traccc_add_executable( seq_example_cuda "seq_example_cuda.cpp"
-   LINK_LIBRARIES vecmem::core vecmem::cuda
-                  traccc::io traccc::algorithms traccc::algorithms_cuda traccc::performance 
-                  Boost::program_options)
+   LINK_LIBRARIES vecmem::core vecmem::cuda traccc::io traccc::performance 
+                  traccc::core traccc::cuda Boost::program_options)
 traccc_add_executable( seeding_example_cuda "seeding_example_cuda.cpp"
-   LINK_LIBRARIES vecmem::core vecmem::cuda
-                  traccc::io traccc::algorithms traccc::algorithms_cuda traccc::performance
-                  Boost::program_options)
+   LINK_LIBRARIES vecmem::core vecmem::cuda traccc::io traccc::performance
+                  traccc::core traccc::cuda Boost::program_options)
