+++ examples/run/cuda/seeding_example_cuda.cpp
@@ -11,10 +11,10 @@
 #include "traccc/io/writer.hpp"
 
 // algorithms
+#include "traccc/cuda/seeding/seeding_algorithm.hpp"
 #include "traccc/cuda/seeding/track_params_estimation.hpp"
-#include "traccc/cuda/track_finding/seeding_algorithm.hpp"
+#include "traccc/seeding/seeding_algorithm.hpp"
 #include "traccc/seeding/track_params_estimation.hpp"
-#include "traccc/track_finding/seeding_algorithm.hpp"
 
 // performance
 #include "traccc/efficiency/seeding_performance_writer.hpp"
