+++ examples/run/sycl/seq_example_sycl.sycl
@@ -16,10 +16,10 @@
 
 // algorithms
 #include "traccc/clusterization/clusterization_algorithm.hpp"
+#include "traccc/seeding/seeding_algorithm.hpp"
 #include "traccc/seeding/track_params_estimation.hpp"
+#include "traccc/sycl/seeding/seeding_algorithm.hpp"
 #include "traccc/sycl/seeding/track_params_estimation.hpp"
-#include "traccc/sycl/track_finding/seeding_algorithm.hpp"
-#include "traccc/track_finding/seeding_algorithm.hpp"
 
 // performance
 #include "traccc/efficiency/seeding_performance_writer.hpp"
